requirements.txt
bcrypt
cryptography
customtkinter
matplotlib
mysql-connector-python
openpyxl
pandas
pillow
plyer
psutil
pynput
pystray
python-dotenv
pywin32
reportlab
tendo
tkcalendar
tkinter-tooltip
keyring
winotify
auto-py-to-exe